Split ReferencedYamlTest for non-osgi and osgi
